Refactor lifecycle utils.rs to address low cohesion
Split the monolithic utils.rs (47 functions across 8 responsibilities) into
focused modules:

- spawning.rs: spawn_daemon, resolve_daemon_binary
- monitoring.rs: wait_for_ready, read_health, read_pid, HealthSnapshot,
  HealthCheckOutcome, snapshot validation
- shutdown.rs: wait_for_shutdown, signal_daemon
- socket.rs: ensure_socket_available, socket_is_reachable, connectivity probing

utils.rs now contains only high-level orchestration:
- ensure_no_extra_arguments, prepare_runtime, write_startup_banner
- try_auto_start_daemon (public API)
- STARTUP_TIMEOUT and AUTO_START_TIMEOUT constants

This refactoring improves cohesion by grouping related functionality and
reduces the file from 749 lines to 234 lines, with each new module under
175 lines.
